refactor : dsl 코드 개선
