+++ app/src/main/java/com/donglab/screennameviewer/SampleApplication.kt
@@ -1,7 +1,9 @@
 package com.donglab.screennameviewer
 
 import android.app.Application
-import com.donglab.screennameviewer.publicapi.viewer.initScreenNameViewer
+import android.graphics.Color
+import android.view.Gravity
+import com.donglab.screennameviewer.publicapi.dsl.initScreenNameViewer
 
 class SampleApplication : Application() {
 
@@ -12,6 +14,22 @@ class SampleApplication : Application() {
             settings {
                 debugMode { true }
                 enabled { true }
+            }
+            config {
+                textStyle {
+                    size = 10f
+                    color = Color.BLUE
+                }
+                background {
+                    color = Color.argb(80, 255, 255, 255)
+                    padding = 20
+                }
+                position {
+                    topMargin = 52
+                    activity = Gravity.TOP or Gravity.START
+                    fragment = Gravity.TOP or Gravity.END
+                    composeRoute = Gravity.TOP or Gravity.END
+                }
             }
         }
     }
