+++ screennameviewer/compose/src/main/java/com/donglab/screennameviewer/publicapi/config/ScreenNameOverlayConfig.kt
@@ -1,7 +1,6 @@
 package com.donglab.screennameviewer.publicapi.config
 
-import android.graphics.Color
-import android.view.Gravity
+import com.donglab.screennameviewer.publicapi.dsl.overlayConfig
 
 /**
  * ScreenNameViewer의 UI 설정을 담당하는 설정 클래스
@@ -17,22 +16,6 @@ data class ScreenNameOverlayConfig internal constructor(
     val composeRouteGravity: Int
 ) {
     companion object {
-
-        fun default(): ScreenNameOverlayConfig = overlayConfig {
-            textStyle {
-                size = 10f
-                color = Color.BLUE
-            }
-            background {
-                color = Color.argb(50, 200, 200, 200)
-                padding = 16
-            }
-            position {
-                topMargin = 52
-                activity = Gravity.TOP or Gravity.START
-                fragment = Gravity.TOP or Gravity.END
-                composeRoute = Gravity.TOP or Gravity.END
-            }
-        }
+        fun default(): ScreenNameOverlayConfig = overlayConfig {}
     }
 }
