+++ screennameviewer/compose/src/main/java/com/donglab/screennameviewer/publicapi/dsl/SettingDsl.kt
@@ -1,46 +1,38 @@
-package com.donglab.screennameviewer.publicapi.setting
+package com.donglab.screennameviewer.publicapi.dsl
+
+import com.donglab.screennameviewer.publicapi.setting.ScreenNameViewerSetting
 
 /**
- * ScreenNameViewerSetting을 위한 DSL Builder
+ * DSL 스코프 마커 - 중첩된 DSL 블록에서 외부 스코프 접근을 방지
  */
+@DslMarker
+annotation class SettingDsl
+
+@SettingDsl
 class SettingBuilder {
     private var debugModeCondition: () -> Boolean = { false }
     private var enabledCondition: () -> Boolean = { false }
 
     /**
-     * 디버그 모드 조건 설정
+     * 디버그 모드 활성화 조건을 람다로 설정하는 확장함수
      */
     fun debugMode(condition: () -> Boolean) {
         debugModeCondition = condition
     }
 
     /**
-     * 활성화 조건 설정
+     * 활성화 조건을 람다로 설정하는 확장함수
      */
     fun enabled(condition: () -> Boolean) {
         enabledCondition = condition
     }
 
-    /**
-     * ScreenNameViewerSetting 인스턴스 생성
-     */
     fun build(): ScreenNameViewerSetting = ScreenNameViewerSetting(
         debugModeCondition = debugModeCondition,
         enabledCondition = enabledCondition
     )
 }
 
-/**
- * DSL 스타일로 ScreenNameViewerSetting을 생성하는 진입점 함수
- *
- * 사용 예시:
- * ```kotlin
- * val setting = setting {
- *     debugMode { BuildConfig.DEBUG }
- *     enabled { true }
- * }
- * ```
- */
-fun setting(block: SettingBuilder.() -> Unit): ScreenNameViewerSetting {
+inline fun setting(block: SettingBuilder.() -> Unit): ScreenNameViewerSetting {
     return SettingBuilder().apply(block).build()
 }
