+++ screennameviewer/compose/src/main/java/com/donglab/screennameviewer/publicapi/setting/ScreenNameViewerSetting.kt
@@ -1,5 +1,7 @@
 package com.donglab.screennameviewer.publicapi.setting
 
+import com.donglab.screennameviewer.publicapi.dsl.setting
+
 /**
  * ScreenNameViewer의 활성화 조건을 정의하는 설정 클래스
  */
@@ -14,7 +16,6 @@ data class ScreenNameViewerSetting internal constructor(
         get() = isDebugMode && enabledCondition()
 
     companion object {
-
         fun default(): ScreenNameViewerSetting = setting {
             debugMode { false }
             enabled { false }
